satellite/accounting: move code and remove unused fields
