+++ satellite/console/usagerollups_test.go
@@ -102,10 +102,8 @@ func TestUsageRollups(t *testing.T) {
 					BucketName:     []byte(bucket),
 					ProjectID:      project1,
 					ObjectCount:    value1,
-					Segments:       value1,
 					InlineSegments: value1,
 					RemoteSegments: value1,
-					Bytes:          value1,
 					InlineBytes:    value1,
 					RemoteBytes:    value1,
 					MetadataSize:   value1,
@@ -115,10 +113,8 @@ func TestUsageRollups(t *testing.T) {
 					BucketName:     []byte(bucket),
 					ProjectID:      project2,
 					ObjectCount:    value2,
-					Segments:       value2,
 					InlineSegments: value2,
 					RemoteSegments: value2,
-					Bytes:          value2,
 					InlineBytes:    value2,
 					RemoteBytes:    value2,
 					MetadataSize:   value2,
